fix: go directly to search location screen when adding favorite (#832)
* fix: go directly to search location screen when adding favorite

Fixes #674

* fix: closes entire modal when saving favorite

* fix: typing issue with default screen

* fix: keyboard avoidance on android

* refactor: removes irrelevant changes

* fixes type issue

* Removes unused import
